Reorder and note (#1077)
